add SciGlassCalorimeterTiltless
This is an updated design for the SciGlass barrel electromagnetic
calorimeter. The main change compared to the
SciGlassCalorimeterTiltless is introduction of tower families and
additional flaring to reduce the gaps between the towers. The most
significant change is the removal of the phi tilt (hence the temporary
name suffix "Tiltless"), which significantly simplifies the
construction of the geometry.